utils/keyboardGenerator/pyproject.toml
[tool.poetry]
name = "keyboardgenerator"
version = "0.1.0"
description = ""
authors = ["Yehonatan Elentok <yehonatan.e@solaredge.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
pykle-serial = {git = "https://github.com/hajimen/pykle_serial.git"}
matplotlib = "^3.8.0"
scipy = "^1.11.3"
solidpython2 = "^2.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
